phase-3-chatbot/requirements.txt
openai>=1.30.0
openai-agents>=0.0.4
httpx>=0.27.0
python-dotenv>=1.0.0
mcp>=1.0.0
